move _check_bits_against_stack to lsst_bits.py
